refactor: move main.ts into app
