+++ embeddable-say-editor/test-app/multiple-editors/multiple-editors.ts
@@ -1,4 +1,4 @@
-import { renderEditor } from '../../main.ts';
+import { renderEditor } from '../../app/main.ts';
 import { router } from '../router.ts';
 
 document.body.appendChild(router);
